Allow the inversion of the file order in albums.
This is primarily useful to keep the filenames stable.
Google Photos by default gives the latest photo the "first" position.
This results in all symlinks renaming if new album photos appear.
With this flag, old pictures will mostly keep their name and new ones
will get higher number prefixes as long as no photos are added in the
middle of a timeline, which is pretty rare for auto-uploaded photos.